sdk/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
openai-agents
